requirements.txt
requests>=2.31.0
pandas>=2.0.0
pyyaml>=6.0
python-telegram-bot>=20.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
html2text>=2020.1.16
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
